+++ lib/ChunkrefMemoryHandler.h
@@ -38,26 +38,26 @@ namespace ChunkRefMemoryHandler {
 
     class UnusedMemorySubchunk_t {
     public:
-        uint16_t                size_;
+        uint32_t                size_;
         UnusedMemorySubchunk_t* nextFree_;
 
-        UnusedMemorySubchunk_t(uint16_t aSize);
-        UnusedMemorySubchunk_t(uint16_t aSize, UnusedMemorySubchunk_t *aNextFree);
+        UnusedMemorySubchunk_t(uint32_t aSize);
+        UnusedMemorySubchunk_t(uint32_t aSize, UnusedMemorySubchunk_t *aNextFree);
         inline byte* deliver();
 
     } __attribute__((packed));
 
 
-    template<uint16_t kSizeChunk, uint8_t kSizeCacheLine, bool kBestFit>
+    template<uint32_t kSizeChunk, uint8_t kSizeCacheLine, bool kBestFit>
     class MemoryChunk_t {
     public:
         MemoryChunk_t();
 
         void freeChunk();
         uint32_t getBytesAllocated();
-        byte *getMem(uint16_t aSize);
+        byte *getMem(uint32_t aSize);
         bool verify();
-        void release(byte *aStartAddr, uint16_t aSize);
+        void release(byte *aStartAddr, uint32_t aSize);
         uint32_t getFree();
         inline bool contains(UnusedMemorySubchunk_t *aAddr);
         inline bool contains(byte *aAddr);
@@ -67,12 +67,12 @@ namespace ChunkRefMemoryHandler {
         byte *begin_;
         UnusedMemorySubchunk_t *firstFree_;
 
-        inline uint16_t roundUp(uint16_t aSize);
-        byte *firstFit(uint16_t aSize);
-        byte *bestFit(uint16_t aSize);
+        inline uint16_t roundUp(uint32_t aSize);
+        byte *firstFit(uint32_t aSize);
+        byte *bestFit(uint32_t aSize);
     };
 
-    template<uint16_t kSizeChunk, uint8_t kSizeCacheLine, bool kBestFit>
+    template<uint32_t kSizeChunk, uint8_t kSizeCacheLine, bool kBestFit>
     class MemoryHandler_t {
     public:
         MemoryHandler_t();
@@ -80,7 +80,7 @@ namespace ChunkRefMemoryHandler {
 
         byte* getMem(uint32_t aSize);
         bool verifyPointers();
-        void release(byte *aStartAddr, uint16_t aSize);
+        void release(byte *aStartAddr, uint32_t aSize);
         std::vector<uint32_t>* getBytesAllocatedPerChunk();
         void printUsage();
         void getUsage(MemUsageStats_t& result);
